[Feat] 리뷰 CRUD 기능

+++ src/main/java/com/umust/dobonglife/domain/course/service/CourseService.java
@@ -17,11 +17,13 @@ import com.umust.dobonglife.domain.course.domain.repository.CourseRepository;
 import com.umust.dobonglife.domain.course.domain.vo.CourseBasicInfo;
 import com.umust.dobonglife.domain.course.domain.vo.CourseOperationInfo;
 import com.umust.dobonglife.domain.course.domain.vo.CoursePolicyInfo;
+import com.umust.dobonglife.domain.courseLike.service.CourseLikeService;
 import com.umust.dobonglife.domain.user.service.UserService;
 import com.umust.dobonglife.global.common.exception.BusinessException;
 import com.umust.dobonglife.global.common.response.CursorResponse;
 import com.umust.dobonglife.global.common.response.ErrorCode;
 import com.umust.dobonglife.global.common.s3.S3Utils;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
@@ -30,15 +32,18 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class CourseService {
     private final CourseRepository courseRepository;
     private final CoursePlansRepository coursePlansRepository;
     private final S3Utils s3Utils;
     private final UserService userService;
+    private final CourseLikeService courseLikeService;
 
     @Transactional(readOnly = true)
     public CursorResponse<CourseSummaryResponse> getCourses(Long lastId, int size) {
@@ -56,20 +61,26 @@ public class CourseService {
     }
 
     @Transactional(readOnly = true)
-    public CourseDetailResponse getCourse(Long courseId) {
+    public CourseDetailResponse getCourse(Long userId, Long courseId) {
+        boolean isRemoved = userService.isCourseRemoved(userId, courseId);
+        boolean isFavorite = courseLikeService.isCourseFavorite(userId, courseId);
+
         Course course = courseRepository.findByIdWithDescription(courseId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.INVALID_COURSE_ID));
 
         List<CoursePlans> plans = coursePlansRepository
                 .findByCourseIdOrderByDateTime(courseId);
 
-        return CourseDetailResponse.from(course, plans);
+        return CourseDetailResponse.from(course, plans, isRemoved, isFavorite);
     }
 
     @Transactional
     public CourseRegisterResponse createCourse(Long userId, CreateCourseRequest request, List<MultipartFile> imageFiles) {
         // 이미지
-        List<String> imageUrls = s3Utils.uploadImages(imageFiles);
+        List<String> imageUrls = new ArrayList<>();
+        if(imageFiles != null && !imageFiles.isEmpty() && !imageFiles.get(0).isEmpty()) {
+            imageUrls = s3Utils.uploadImages(imageFiles);
+        }
 
         // vo
         CourseBasicInfo basicInfo = new CourseBasicInfo(request.title(), request.subTitle(), request.duration(), CourseLevel.valueOf(request.level()));
@@ -138,14 +149,14 @@ public class CourseService {
     }
 
     private void updateImageUrls(Course course, List<String> urlsToDelete, List<MultipartFile> newFiles) {
-        if (urlsToDelete != null) {
+        if (urlsToDelete != null && !urlsToDelete.isEmpty() && !urlsToDelete.get(0).isEmpty()) {
             urlsToDelete.forEach(url -> {
                 s3Utils.deleteImage(url);
                 course.getImageUrls().remove(url);
             });
         }
 
-        if (newFiles != null && !newFiles.isEmpty()) {
+        if (newFiles != null && !newFiles.isEmpty() && !newFiles.get(0).isEmpty()) {
             List<String> images = s3Utils.uploadImages(newFiles);
             course.getImageUrls().addAll(images);
         }
